Split models and criterion and fixed tests import
