+++ sparse_ho/__init__.py
@@ -1,7 +1,7 @@
 from .ho import grad_search
 from .ho import grad_search_wolfe
-from .backward import Backward
-from .forward import Forward
-from .implicit_forward import ImplicitForward
+from .algo.backward import Backward
+from .algo.forward import Forward
+from .algo.implicit_forward import ImplicitForward
 
 __version__ = '0.1.dev'
