+++ sparse_ho/tests/test_elastic.py
@@ -1,13 +1,13 @@
 import numpy as np
 from sklearn import linear_model
-from sparse_ho.models import ElasticNet
-from sparse_ho.forward import get_beta_jac_iterdiff
+from sparse_ho.model.enet import ElasticNet
+from sparse_ho.algo.forward import get_beta_jac_iterdiff
 from sparse_ho.datasets.synthetic import get_synt_data
-from sparse_ho.implicit_forward import get_beta_jac_fast_iterdiff
-from sparse_ho.criterion import HeldOutMSE
-from sparse_ho.forward import Forward
-from sparse_ho.implicit import Implicit
-from sparse_ho.implicit_forward import ImplicitForward
+from sparse_ho.algo.implicit_forward import get_beta_jac_fast_iterdiff
+from sparse_ho.criterion.hout_mse import HeldOutMSE
+from sparse_ho.algo.forward import Forward
+from sparse_ho.algo.implicit import Implicit
+from sparse_ho.algo.implicit_forward import ImplicitForward
 from sparse_ho.ho import grad_search
 from sparse_ho.utils import Monitor
 from scipy.sparse import csc_matrix
