+++ sparse_ho/tests/test_grad_search.py
@@ -6,12 +6,13 @@ from sklearn import linear_model
 from sparse_ho.utils import Monitor
 
 from sparse_ho.datasets.synthetic import get_synt_data
-from sparse_ho.models import Lasso
+from sparse_ho.model.lasso import Lasso
 
-from sparse_ho.forward import Forward
-from sparse_ho.implicit_forward import ImplicitForward
-from sparse_ho.implicit import Implicit
-from sparse_ho.criterion import HeldOutMSE, SmoothedSURE
+from sparse_ho.algo.forward import Forward
+from sparse_ho.algo.implicit_forward import ImplicitForward
+from sparse_ho.algo.implicit import Implicit
+from sparse_ho.criterion.hout_mse import HeldOutMSE
+from sparse_ho.criterion.smoothed_sure import SmoothedSURE
 from sparse_ho.ho import grad_search
