+++ sparse_ho/tests/test_lasso.py
@@ -4,16 +4,18 @@ from sklearn import linear_model
 
 
 from sparse_ho.datasets.synthetic import get_synt_data
-from sparse_ho.forward import get_beta_jac_iterdiff
-from sparse_ho.models import Lasso, WeightedLasso
-from sparse_ho.implicit_forward import get_beta_jac_fast_iterdiff
-from sparse_ho.implicit import get_beta_jac_t_v_implicit
+from sparse_ho.algo.forward import get_beta_jac_iterdiff
+from sparse_ho.model.lasso import Lasso
+from sparse_ho.model.wlasso import WeightedLasso
+from sparse_ho.algo.implicit_forward import get_beta_jac_fast_iterdiff
+from sparse_ho.algo.implicit import get_beta_jac_t_v_implicit
 
-from sparse_ho.forward import Forward
-from sparse_ho.implicit_forward import ImplicitForward
-from sparse_ho.implicit import Implicit
-from sparse_ho.backward import Backward
-from sparse_ho.criterion import HeldOutMSE, SmoothedSURE
+from sparse_ho.algo.forward import Forward
+from sparse_ho.algo.implicit_forward import ImplicitForward
+from sparse_ho.algo.implicit import Implicit
+from sparse_ho.algo.backward import Backward
+from sparse_ho.criterion.hout_mse import HeldOutMSE
+from sparse_ho.criterion.smoothed_sure import SmoothedSURE
 
 n_samples = 100
 n_features = 100
