+++ sparse_ho/tests/test_logreg.py
@@ -4,13 +4,13 @@ from sklearn import datasets
 from sklearn.linear_model import LogisticRegression
 from scipy.sparse import csc_matrix
 
-from sparse_ho.models import SparseLogreg
-from sparse_ho.forward import get_beta_jac_iterdiff
-from sparse_ho.implicit_forward import get_beta_jac_fast_iterdiff
-from sparse_ho.forward import Forward
-from sparse_ho.implicit_forward import ImplicitForward
-from sparse_ho.implicit import Implicit
-from sparse_ho.criterion import HeldOutLogistic
+from sparse_ho.model.sparselogreg import SparseLogreg
+from sparse_ho.algo.forward import get_beta_jac_iterdiff
+from sparse_ho.algo.implicit_forward import get_beta_jac_fast_iterdiff
+from sparse_ho.algo.forward import Forward
+from sparse_ho.algo.implicit_forward import ImplicitForward
+from sparse_ho.algo.implicit import Implicit
+from sparse_ho.criterion.hout_logistic import HeldOutLogistic
 from sparse_ho.utils import Monitor
 from sparse_ho.ho import grad_search
