+++ sparse_ho/tests/test_svm.py
@@ -2,16 +2,16 @@ import numpy as np
 import pytest
 from sklearn import datasets
 from sklearn.svm import LinearSVC
-from sparse_ho.criterion import HeldOutLogistic
-from sparse_ho.forward import Forward
-from sparse_ho.implicit import Implicit
-from sparse_ho.implicit_forward import ImplicitForward
-from sparse_ho.models import SVM
-from sparse_ho.forward import get_beta_jac_iterdiff
-from sparse_ho.implicit_forward import get_beta_jac_fast_iterdiff
+from sparse_ho.criterion.hout_logistic import HeldOutLogistic
+from sparse_ho.algo.forward import Forward
+from sparse_ho.algo.implicit import Implicit
+from sparse_ho.algo.implicit_forward import ImplicitForward
+from sparse_ho.model.svm import SVM
+from sparse_ho.algo.forward import get_beta_jac_iterdiff
+from sparse_ho.algo.implicit_forward import get_beta_jac_fast_iterdiff
 from scipy.sparse import issparse
 from sparse_ho.ho import grad_search
-from sparse_ho.criterion import HeldOutSmoothedHinge
+from sparse_ho.criterion.hout_smoothed_hinge import HeldOutSmoothedHinge
 from sparse_ho.utils import Monitor
